requirements.txt
asyncpraw==7.7.0
beautifulsoup4==4.12.2
# controls==0.0.0
fastapi==0.95.1
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
gTTS==2.3.1
jose==1.0.0
Markdown==3.4.3
moviepy==1.0.3
openai==0.27.4
pandas==1.5.3
passlib==1.7.4
Pillow==9.5.0
praw==7.7.0
prawcore==2.3.0
pydantic==1.10.6
python-dotenv==1.0.0
python_jose==3.3.0
pyttsx3==2.90
requests==2.28.2
selenium==4.9.0
SQLAlchemy==2.0.7
uvicorn==0.21.1
psycopg2==2.9.6